requirements.txt
redis<=2.10.6
aiohttp
schedule
pyquery
requests
flask
sanic;sys_platform!='win32'
qqwry-py3
